backend/requirements.txt
asgiref==3.8.1
Django==5.2.2
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
joblib==1.5.1
numpy==2.2.6
scikit-learn==1.6.1
scipy==1.15.3
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
